Added the `contain` method for `Chromosome`s.
This method searches for a `E*` with matching role(s), state(s), and
bounding identification tags. It is useful for identifying whether a
`E` is already encoded in a given `Genotype`.

`Genotype` now employs this method in its `add_link` and `add_node`
methods. Its usage prevents these mutation operators from adding repeated
or ambiguous elements to the `Genotype`.