Simplified the installation script and updated the README markdown file.
It turns out many of the checks and commands present in the previous
version of 'install.sh' are automatically performed and run by CMake.
The revised installation script takes that into consideration, and is
therefore much simpler.

The 'README.md' file has been updated to make the installation
instructions system-agnostic and also valid for users who do not
possess root privileges.