Parallelization and some clean-up.
The code is now fully paralellized at the C++ level. As a result, Python
scripts no longer need to bother with MPI as that is now automatically
handled.

The Novelty Search mode has also been removed as its implementation was
hindering the parallelization of the code. It will be re-designed and
re-implemented in the future.

The random nature of the code has now also been made explicit by the
`random_seed` parameter. This parameter ensures that all nodes running
the code in parallel will have matching evolutionary histories.

Other parameters have been removed, and the documentation has been
expanded and updated to reflect all the changes listed above.