Allows additional flexibility in `Experiment` operations.
Now the user can also have the experiment execute other operations
at each cycle which are not necessarily the computation of a
fitness score or the display of information on the screen.

The `execute` method can be overridden from Python to make an
`Experiment` instance perform any operation which might be needed.